final minor tweaks
