+++ sample.py
@@ -14,7 +14,7 @@ from diffusers.models import AutoencoderKL
 
 from modules.diffusion import create_diffusion
 from download import find_model
-from models import DiT_models
+from modules.dit_builder import DiT_models
 
 torch.backends.cuda.matmul.allow_tf32 = True
 torch.backends.cudnn.allow_tf32 = True
